move plotting utility functions to separate file
